Add flights just signed up for to upcoming flight list
Added code to track the flights an attendant has just signed up for and add those to the "upcoming flights" list. There is a lag between sign-up and the event getting added to the attendant's calendar. This caused an unexpected experience for the user: they would sign up for a flight, the page would refresh, and the flight would be gone from "Available Flights" but not be present in "Upcoming Flights".